add votes and password routes
